Update kimber.md - remove excerpt tag from frontmatter
There's a bug in Jekyll version 3.9.4 (recently released), something to do with the Excerpt tag in the frontmatter (metadata).

According to the workaround, removing the Excerpt tag from the 404.md, about.md, and non-menu-page.md should work. The error in my first build points to a problem with 404.md (see below), but after fixing that the error went to allyship.md. I think the Excerpt tag needs to be removed from all files under the _pages directory.

I removed it from several of the .md files, but it's still present in a few of them and the build is breaking on those.

Bug details: jekyll/jekyll#9544

Workaround: academicpages/academicpages.github.io#1878

Error in the build file first time:
/usr/local/bundle/gems/jekyll-3.9.4/lib/jekyll/excerpt.rb:135:in extract_excerpt': undefined method excerpt_separator' for #<Jekyll::Page @name="404.md"> (NoMethodError)

Error in the build file second time:
/usr/local/bundle/gems/jekyll-3.9.4/lib/jekyll/excerpt.rb:135:in extract_excerpt': undefined method excerpt_separator' for #<Jekyll::Page @name="allyship.md"> (NoMethodError)

Error in the build file fourth time:
/usr/local/bundle/gems/jekyll-3.9.4/lib/jekyll/excerpt.rb:135:in `extract_excerpt': undefined method `excerpt_separator' for #<Jekyll::Page @name="kimber.md"> (NoMethodError)